MASTER: Everything implemented appears to be functioning
All Elf_Details functions appear to be working as expected.
Bad News:  I'm not using a Debugger.
Good News:  I programmed in almost the maximum amount of error checking/handling known to man

Elven_Chain appears to be fully functional as well but this is less impressive as Elven_Chain is merely the glue that runs main().
-Hark