MASTER: Added member to Elf_Detail struct
New member is the char* magicNum.  It will hold the first four bytes of the file contents if they match the ELF magic number.  This enables the print function to entirely ignore the struct if magicNum doesn't match the ELF magic number.  Checked during Input Validation of the print function, added a stderr statement to inform the user.  New struct member required update of:
    	* parse_elf()
	* print_elf_details()
	* kill_elf()

README notes for integrating a new struct member appear to be complete.  -Hark