MASTER: Implemented convert_char_to_int()
This function converts a certain number of bytes into an unsigned int based on the given endianess of the system.

"So let it be written.  So let it be [tested]."
"Creeping Death" Metallica

I'm pretty sure that's what Metallica meant.  Regardless, I feel a compulsion to actually test this function.  It will be the first real test so the next step is to create a test folder and write some test code.  -Hark