MASTER: Implemented index to Section Header entry with Names
This implements everything listed in the ELF Header (not counting the Program Header, Section Header, etc).  Also, added some ideas to To Do (in Readme) to validate some of the information we discovered in the ELF Header.  Now, it's time to do some comparison testing, release the first version, branch, and then get to work on the next section... the Program Header. -Hark