PRGRM_HDR: Necessary refactor discovered
Incremental, untested, offline commit.  I just discovered that I want to refactor this branch to alleviate the need for the PH parsing functions to get the Elf Header struct.  All the necessary information should be duplicated into the PH Details struct when it is allocated and initialized inside parse_program_header().  -Hark