PRGRM_HDR: Correction regarding print_binary()
print_binary() was right all along.  (I thought that was fishy)  Turns out I was passing in the wrong endianness for this development system as a hard-coded value all along.  Now that I passed in the actuall endianness from the program header details struct (*NOT* the segment structs) all is right with the world. -Hark